libdisni: Remove duplicate define PACKAGE_NAME
The autoconf tool by default defines PACKAGE_NAME to the name of the
overall package, so this reserved name should not be reused elsewhere.
The PACKAGE_NAME define in the nvmef JNI source is not used anywhere in
the file nor does it appear to be required by JNI. Eliminate it.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>