libdisni: Add @dpdk@/include to include path
The "correct" location to look for the DPDK include files is
${RTE_SDK}/${RTE_TARGET}/include; this is where the install scripts and
Linux distributions create symlinks under /usr/share/dpdk/${RTE_TARGET}
which preserve this structure.

While Linux distribution packages place the DPDK header files into the
standard location /usr/include/dpdk, adapting the configure script to
automatically add ${includedir}/dpdk and then check if the header is
there opens a huge can of worms, so this patch doesn't do that.

To not break existing build scripts, @dpdk@/include/dpdk remains in the
include path; if this directory is not present, it does no harm.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>