libdisni: Do not create static library
There is no reason for end users to generate a static library for a JNI
library, since the JVM can only use the shared library. The static
library can still be enabled with --enable-static in the rare case that
it is needed.

Additionally, modernize the syntax from the deprecated AC_PROG_LIBTOOL
to the modern LT_INIT, which is necessary to make this work.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>