CM Processor: change to daemon thread
Some application (e.g. Spark) rely on the application exiting
when the main method returns. This requires all threads still
active at the return of main to be daemon threads otherwise
the application will wait for the threads to exit. This fix
changes the RdmaCmProcessor thread to be a daemon thread.

Signed-off-by: Jonas Pfefferle <pepperjo@japf.ch>